added CRUD features in branch, area and member of bar assocication and enhance the UI of register page

+++ record_division.py
@@ -37,11 +37,40 @@ def edit_division(id):
     cur.close()
     return render_template('edit_division.html', division=division)
 
+# @record_division_bp.route('/delete/<int:id>', methods=['POST'])
+# def delete_division(id):
+#     cur = mysql.connection.cursor()
+#     cur.execute("DELETE FROM tbl_division WHERE division_id=%s", (id,))
+#     mysql.connection.commit()
+#     cur.close()
+#     flash('Division deleted successfully!', 'success')
+#     return redirect(url_for('record_division.record_division'))
+
 @record_division_bp.route('/delete/<int:id>', methods=['POST'])
 def delete_division(id):
     cur = mysql.connection.cursor()
+    # Check if any branches or areas use this division
+    cur.execute("SELECT branch_name FROM tbl_branch WHERE division_id=%s", (id,))
+    blocking_branches = cur.fetchall()
+    cur.execute("SELECT area_name FROM tbl_area WHERE division_id=%s", (id,))
+    blocking_areas = cur.fetchall()
+
+    if blocking_branches or blocking_areas:
+        msg = []
+        if blocking_branches:
+            branch_list = ', '.join([b[0] for b in blocking_branches])
+            msg.append(f'Branches: {branch_list}')
+        if blocking_areas:
+            area_list = ', '.join([a[0] for a in blocking_areas])
+            msg.append(f'Areas: {area_list}')
+        flash(f'Cannot delete division. In use by {", ".join(msg)}', 'danger')
+        cur.close()
+        return redirect(url_for('record_division.record_division'))
+
+    # Safe to delete
     cur.execute("DELETE FROM tbl_division WHERE division_id=%s", (id,))
     mysql.connection.commit()
     cur.close()
     flash('Division deleted successfully!', 'success')
     return redirect(url_for('record_division.record_division'))
+
